# apis/Ejercicio1_Libro/swagger.yaml
openapi: 3.0.3
info:
  title: API de Gestión de Biblioteca
  description: API RESTful para la administración de libros en una biblioteca.
  version: 1.0.0
paths:
  /api/v1/libros:
    post:
      summary: Registrar un nuevo libro
      operationId: registrarLibro
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LibroRequest'
      responses:
        '201':
          description: Libro registrado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Libro'
        '400':
          description: Solicitud incorrecta o datos inválidos

    get:
      summary: Consultar todos los libros
      operationId: consultarLibros
      responses:
        '200':
          description: Lista de libros obtenida exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Libro'

  /api/v1/libros/buscar:
    get:
      summary: Consultar libro por título
      operationId: consultarLibroPorTitulo
      parameters:
        - name: titulo
          in: query
          required: true
          description: Título o parte del título del libro a buscar
          schema:
            type: string
      responses:
        '200':
          description: Coincidencias encontradas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Libro'
        '400':
          description: Parámetro de búsqueda no proporcionado

  /api/v1/libros/{id}:
    get:
      summary: Consultar libro por ID
      operationId: consultarLibroPorId
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Libro encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Libro'
        '404':
          description: Libro no encontrado

    put:
      summary: Actualizar un libro existente
      operationId: actualizarLibro
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LibroRequest'
      responses:
        '200':
          description: Libro actualizado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Libro'
        '400':
          description: Datos de entrada inválidos
        '404':
          description: Libro no encontrado

    delete:
      summary: Eliminar un libro
      operationId: eliminarLibro
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Libro eliminado exitosamente (sin contenido en respuesta)
        '404':
          description: Libro no encontrado

components:
  schemas:
    Libro:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        titulo:
          type: string
          example: "Cien años de soledad"
        autor:
          type: string
          example: "Gabriel García Márquez"
        isbn:
          type: string
          example: "978-0307474728"
        añopub:
          type: integer
          example: 1967
        estado:
          type: string
          example: "DISPONIBLE"

    LibroRequest:
      type: object
      required:
        - titulo
        - autor
        - isbn
        - añopub
        - estado
      properties:
        titulo:
          type: string
          example: "Cien años de soledad"
        autor:
          type: string
          example: "Gabriel García Márquez"
        isbn:
          type: string
          example: "978-0307474728"
        añopub:
          type: integer
          example: 1967
        estado:
          type: string
          example: "DISPONIBLE"